+++ resources/core/boundary/adapter/persistence/connector.clj
@@ -1,0 +1,10 @@
+;; This Source Code Form is subject to the terms of the Mozilla Public
+;; License, v. 2.0. If a copy of the MPL was not distributed with this
+;; file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+{{=<< >>=}}
+(ns <<namespace>>.boundary.adapter.persistence.connector)
+
+<<#hydrogen-persistence-sql?>>
+(defrecord Sql [db-spec])
+<</hydrogen-persistence-sql?>>